requirements.txt
click~=7.1.2
pyserial~=3.5
PyYAML==6.0.2
websockets==15.0.1
aiohttp==3.11.18